backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
sentence-transformers==3.0.1
faiss-cpu==1.8.0
PyMuPDF==1.24.5
langchain==0.2.12
langchain-community==0.2.11
transformers==4.43.3
torch==2.3.1
requests==2.32.3
numpy==1.26.4
werkzeug==3.0.3
